feat: new worker registry page added to project
feat: nav-bar items are only visible for those users who are allowed … 
feat: path to registry page added
feat: pages redirect users who are not allowed to be there
style: added styling to registery page
feat: business logic added to registry page in form

+++ core_webshop/phoneshop/views.py
@@ -1,11 +1,15 @@
+from django.utils import timezone
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
 from django.views.decorators.http import require_POST
 from django.contrib.auth import authenticate, login as auth_login
 from django.http import HttpResponseBadRequest
 from django.shortcuts import render, redirect, get_object_or_404
-from .models import Products, Cart, Shops, Users, Sales
+from .models import Products, Cart, Shops, Users, Sales, Workers
 from support_files.sorting import sort_product
 from django.http import HttpResponse, JsonResponse
+from support_files.register import RegistrationForm
+
 
 @login_required(login_url='/')
 def index(request):
@@ -28,10 +32,14 @@ def index(request):
 
 @login_required(login_url='/')
 def cart(request):
+    if request.user.is_superuser:
+        return redirect('home')
     return render(request, 'cart.html')
 
 @require_POST
 def delete_cart_item(request, item_id):
+    if request.user.is_superuser:
+        return redirect('home')
     try:
         item = Cart.objects.get(id=item_id)
         item.delete()
@@ -56,12 +64,16 @@ def product_detail(request, name):
 
 @login_required
 def user_cart(request):
+    if request.user.is_superuser:
+        return redirect('home')
     phoneshop_user = request.user.phoneshop_user
     cart_items = Cart.objects.filter(user_id=phoneshop_user.id)
     return render(request, 'cart.html', {'cart_items': cart_items})
 
 @login_required
 def add_to_cart(request, product_id):
+    if request.user.is_superuser:
+        return redirect('home')
     phoneshop_user = request.user.phoneshop_user
     print("Adding to cart for:", phoneshop_user.id)
     product = get_object_or_404(Products, id=product_id)
@@ -97,3 +109,44 @@ def login(request):
         else:
             return render(request, 'login.html', {'error': 'Hibás felhasználónév vagy jelszó!'})
     return render(request, 'login.html')
+
+@login_required
+def register(request):
+    if not request.user.is_superuser:
+        return redirect('home')
+    if request.method == "POST":
+        form = RegistrationForm(request.POST)
+        if form.is_valid():
+
+            worker = Workers.objects.create(
+                last_name=form.cleaned_data['last_name'],
+                first_name=form.cleaned_data['first_name'],
+                address=form.cleaned_data['address'],
+                birth_date=form.cleaned_data['birth_date'],
+                phone_number=form.cleaned_data['phone_number'],
+                position=form.cleaned_data['position'],
+                shop=form.cleaned_data['shop'],
+                admin=False
+            )
+
+
+            user = User.objects.create_user(
+                last_name=form.cleaned_data['last_name'],
+                first_name=form.cleaned_data['first_name'],
+                username=form.cleaned_data['username'],
+                email=form.cleaned_data['email'],
+                password=form.cleaned_data['password']
+            )
+
+
+            Users.objects.create(
+                user=user,
+                worker=worker
+            )
+
+            return redirect('home')
+
+    else:
+        form = RegistrationForm()
+
+    return render(request, 'register_worker.html', {'form': form})
